requirements.txt
numpy>=1.21.0
matplotlib>=3.4.0
pillow>=8.2.0
torch>=1.8.0
torchvision>=0.9.0
albumentations>=1.0.0
wandb>=0.12.0
scikit-learn>=0.24.2